feat(query): share completion metadata across API and MCP
Move query completion candidate generation out of the CLI shell helper and into the archive query contract layer so shell, CLI JSON, Python API, and MCP consumers all expose the same payload shape. Relocate public action contracts under operations while keeping the CLI import as a compatibility adapter.

Add facade, MCP, discovery, envelope, and CLI adapter tests around the shared completion payload. Refresh generated topology for the new query contract module.

Verification: nix develop --command devtools verify --quick; focused CLI/API/MCP tests for query completions; MCP read-tool discovery smoke; action-contract compatibility tests. Default devtools verify was attempted but the testmon pytest phase was interrupted after roughly 20 minutes without completing.
