+++ docs/topology-status.md
@@ -19,7 +19,7 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-phase | — | 2 | 2 | 0 | 0 |
 | archive-projection | — | 5 | 5 | 0 | 0 |
 | archive-provider | — | 2 | 2 | 0 | 0 |
-| archive-query | archive query semantics | 22 | 22 | 0 | 0 |
+| archive-query | archive query semantics | 23 | 23 | 0 | 0 |
 | archive-raw-payload | — | 5 | 5 | 0 | 0 |
 | archive-semantic | — | 13 | 13 | 0 | 0 |
 | archive-session | — | 16 | 16 | 0 | 0 |
@@ -28,10 +28,10 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 ### Summary
 
-- **Stable** (no move scoped): 694
+- **Stable** (no move scoped): 695
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (storage-root): 16
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 832
-- **Realized polylogue/**/*.py**: 832 files declared
+- **Total declared**: 834
+- **Realized polylogue/**/*.py**: 834 files declared
 <!-- END GENERATED: topology-status -->
