+++ tests/infra/mcp.py
@@ -48,6 +48,7 @@ EXPECTED_TOOL_NAMES = {
     "get_stats_by",
     "list_read_view_profiles",
     "explain_query_expression",
+    "query_completions",
     "readiness_check",
     "rebuild_index",
     "update_index",
@@ -196,6 +197,7 @@ def make_polylogue_mock(*, resolved_id: str | None = None) -> MagicMock:
     poly.get_logical_session = AsyncMock(return_value=None)
     poly.list_read_view_profiles = AsyncMock(return_value=[])
     poly.explain_query_expression = AsyncMock(return_value={})
+    poly.query_completions = AsyncMock(return_value={})
     poly.get_messages_paginated = AsyncMock(return_value=([], 0))
     poly.get_session = AsyncMock(return_value=None)
     poly.get_session_profile_insight = AsyncMock(return_value=None)
