+++ tests/unit/mcp/test_envelope_contracts.py
@@ -62,6 +62,7 @@ TOOL_CONTRACT: dict[str, ToolKind] = {
     "get_messages": ("envelope", frozenset({"messages", "total"})),
     "list_read_view_profiles": ("envelope", frozenset({"read_views", "total"})),
     "explain_query_expression": "single_object",
+    "query_completions": "single_object",
     "raw_artifacts": ("envelope", frozenset({"raw_artifacts", "total"})),
     "archive_list_sessions": ("envelope", frozenset({"items", "total", "limit", "offset"})),
     "archive_search_sessions": ("envelope", frozenset({"items", "total", "limit", "query"})),
