+++ tests/unit/mcp/test_tool_discovery.py
@@ -69,6 +69,9 @@ _KNOWN_MINIMAL: dict[str, dict[str, object]] = {
     "find_stuck_sessions": {"limit": 1},
     "find_abandoned_sessions": {"limit": 1},
     "get_stats_by": {"group_by": "origin"},
+    "list_read_view_profiles": {},
+    "explain_query_expression": {"expression": "repo:polylogue"},
+    "query_completions": {"kind": "field", "incomplete": "d"},
     "embedding_status": {},
     "embedding_preflight": {},
     "build_context_pack": {"project_repo": "test"},
